requirements.txt
paho-mqtt==1.5.1
black
isort	
interrogate
pdoc3
